Add receipt admin controls Admins can now close receipts, transfer items from closed receipts to the current active receipt, and transfer receipts between models of the same type. This is necessary to keep our purchasing reporting accurate when badges are transferred by admins or in other edge cases.
Technically, this is missing the use-case where someone wants to gift a single item from their receipt to another person. Because there's no good way to split transactions between receipts, and receipt items should not reference transactions on other receipts, this is considered out of scope for the time being.