Finish backend validation redesign With this redesign, we now have a holistic and easy way to define all validations for a set of forms. These validations work with event plugins and also set the correct flags and render keywords. Also, Hayley found that my problem with the x-show fields was that it wasn't contained inside an x-data div.
This still needs some cleanup and for the rest of the WTForms forms to be brought up to speed. I'm also not completely thrilled with the clunky way the overrides and validations get set up using the initialized flag. Other than that, this new system should be able to handle anything we want to throw at it.