Add handling for placeholder validations Sets up the logic for attendee validations to be skipped for placeholders and unassigned badges, and moves the PersonalInfo validations to the new system.
Note that, since we're getting rid of optional_fields (which was very finnicky and hard to work with), the address form no longer carries its own validations. These have been moved to validations/__init__.py alongside the common email and cellphone validations, and must be explicitly enabled for each form. This is a little less convenient than before, but there's a limit to how much code I'm willing to write to keep people from having to explicitly define things in code.