Update Rock Island checklist step for 2025 Adds an explicit dropdown for check-in and check-out times, removes the Quantity field for inventory items, and replaces the per-inventory-item Sample Tracks field with a per-guest Sample Tracks field.
Hopefully the nginx limit on our servers is 100MB because that's what I went with for upload limit size. If it's smaller, people uploading a lot of tracks will run into an ugly error message. We should in the future figure out a better way to handle file uploads and downloads in a consistent fashion.